implementation of article handling (importing mdx files, converting them with mdx-remote, and displaying the content)

+++ src/utils/storage-service.ts
@@ -1,10 +1,15 @@
 'client only';
 
+import { isBrowser } from '@/utils/window';
+
 // storage keys
 export const DARK_MODE_KEY = 'dark_mode';
 
 class StorageService {
   static get<T>(key: string): T | null {
+    if (!isBrowser()) {
+      return null;
+    }
     const item = window.localStorage.getItem(key);
     if (!item) {
       return null;
@@ -21,25 +26,34 @@ class StorageService {
   }
 
   static set<T>(key: string, value: T): void {
-    try {
-      window.localStorage.setItem(key, JSON.stringify(value));
-    } catch (error) {
-      console.error(
-        `Error setting data in localStorage for key ${key}:`,
-        error,
-      );
+    if (isBrowser()) {
+      try {
+        window.localStorage.setItem(key, JSON.stringify(value));
+      } catch (error) {
+        console.error(
+          `Error setting data in localStorage for key ${key}:`,
+          error,
+        );
+      }
     }
   }
 
   static delete(key: string): void {
-    window.localStorage.removeItem(key);
+    if (isBrowser()) {
+      window.localStorage.removeItem(key);
+    }
   }
 
   static clear(): void {
-    window.localStorage.clear();
+    if (isBrowser()) {
+      window.localStorage.clear();
+    }
   }
 
   static exists(key: string): boolean {
+    if (!isBrowser()) {
+      return false;
+    }
     return window.localStorage.getItem(key) !== null;
   }
 }
